requirements.txt
aiosqlite==0.21.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apprise==1.9.5
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.4.0
autograd==1.8.0
cachetools==6.2.1
catboost==1.2.8
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
coolname==2.2.0
cryptography==46.0.3
cycler==0.12.1
dateparser==1.2.2
dill==0.4.0
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.121.0
fonttools==4.60.1
fsspec==2025.10.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hdbscan==0.8.40
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Lifetimes==0.11.3
llvmlite==0.45.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
narwhals==2.10.1
numba==0.62.1
numpy==2.3.4
oauthlib==3.3.1
opentelemetry-api==1.38.0
optuna==4.5.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pendulum==3.1.0
pillow==12.0.0
plotly==6.3.1
pluggy==1.6.0
prefect==3.5.0
prometheus_client==0.23.1
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
readchar==4.2.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.16.3
semver==3.0.4
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
urllib3==2.5.0
uv==0.9.7
uvicorn==0.38.0
websockets==15.0.1
xgboost==3.1.1
zipp==3.23.0
